+++ temba/channels/types/instagram/type.py
@@ -23,7 +23,6 @@ class InstagramType(ChannelType):
 
     courier_url = r"^ig/receive"
     schemes = [URN.INSTAGRAM_SCHEME]
-    redact_values = (settings.FACEBOOK_APPLICATION_SECRET, settings.FACEBOOK_WEBHOOK_SECRET)
 
     claim_blurb = _("Add an %(link)s bot to send and receive messages on behalf of a business Instagram account.") % {
         "link": '<a target="_blank" href="http://instagram.com">Instagram</a>',
@@ -48,6 +47,12 @@ class InstagramType(ChannelType):
             f"https://graph.facebook.com/v18.0/{channel.address}/subscribed_apps",
             params={"access_token": config[Channel.CONFIG_AUTH_TOKEN]},
         )
+
+    def get_redact_values(self, channel) -> tuple:  # pragma: needs cover
+        """
+        Gets the values to redact from logs
+        """
+        return (settings.FACEBOOK_APPLICATION_SECRET, settings.FACEBOOK_WEBHOOK_SECRET)
 
     def get_error_ref_url(self, channel, code: str) -> str:
         return "https://developers.facebook.com/docs/instagram-api/reference/error-codes"