+++ temba/channels/types/whatsapp/tests.py
@@ -7,7 +7,6 @@ from django.test import override_settings
 from django.urls import reverse
 
 from temba.request_logs.models import HTTPLog
-from temba.templates.models import TemplateTranslation
 from temba.tests import MockResponse, TembaTest
 from temba.utils.views import TEMBA_MENU_SELECTION
 
@@ -576,10 +575,7 @@ class WhatsAppTypeTest(TembaTest):
 
     @override_settings(WHATSAPP_ADMIN_SYSTEM_USER_TOKEN="WA_ADMIN_TOKEN")
     @patch("requests.get")
-    def test_get_api_templates(self, mock_get):
-        TemplateTranslation.objects.all().delete()
-        Channel.objects.all().delete()
-
+    def test_fetch_templates(self, mock_get):
         channel = self.create_channel(
             "WAC",
             "WABA name",
@@ -592,29 +588,39 @@ class WhatsAppTypeTest(TembaTest):
         mock_get.side_effect = [
             RequestException("Network is unreachable", response=MockResponse(100, "")),
             MockResponse(400, '{ "meta": { "success": false } }'),
-            MockResponse(200, '{"data": ["foo", "bar"]}'),
+            MockResponse(200, '{"data": ["foo", "bar"]}', headers={"Authorization": "Bearer WA_ADMIN_TOKEN"}),
             MockResponse(
                 200,
                 '{"data": ["foo"], "paging": {"cursors": {"after": "MjQZD"}, "next": "https://graph.facebook.com/v18.0/111111111111111/message_templates?after=MjQZD" } }',
+                headers={"Authorization": "Bearer WA_ADMIN_TOKEN"},
             ),
-            MockResponse(200, '{"data": ["bar"], "paging": {"cursors": {"after": "MjQZD"} } }'),
+            MockResponse(
+                200,
+                '{"data": ["bar"], "paging": {"cursors": {"after": "MjQZD"} } }',
+                headers={"Authorization": "Bearer WA_ADMIN_TOKEN"},
+            ),
         ]
 
-        # RequestException check HTTPLog
-        templates_data, no_error = WhatsAppType().get_api_templates(channel)
-        self.assertEqual(1, HTTPLog.objects.filter(log_type=HTTPLog.WHATSAPP_TEMPLATES_SYNCED).count())
-        self.assertFalse(no_error)
-        self.assertEqual([], templates_data)
+        with self.assertRaises(RequestException):
+            channel.type.fetch_templates(channel)
 
-        # should be empty list with an error flag if fail with API
-        templates_data, no_error = WhatsAppType().get_api_templates(channel)
-        self.assertFalse(no_error)
-        self.assertEqual([], templates_data)
+        self.assertEqual(1, HTTPLog.objects.filter(log_type=HTTPLog.WHATSAPP_TEMPLATES_SYNCED, is_error=True).count())
 
-        # success no error and list
-        templates_data, no_error = WhatsAppType().get_api_templates(channel)
-        self.assertTrue(no_error)
-        self.assertEqual(["foo", "bar"], templates_data)
+        with self.assertRaises(RequestException):
+            channel.type.fetch_templates(channel)
+
+        self.assertEqual(2, HTTPLog.objects.filter(log_type=HTTPLog.WHATSAPP_TEMPLATES_SYNCED, is_error=True).count())
+
+        # check when no next page
+        templates = channel.type.fetch_templates(channel)
+        self.assertEqual(["foo", "bar"], templates)
+
+        self.assertEqual(2, HTTPLog.objects.filter(log_type=HTTPLog.WHATSAPP_TEMPLATES_SYNCED, is_error=True).count())
+        self.assertEqual(1, HTTPLog.objects.filter(log_type=HTTPLog.WHATSAPP_TEMPLATES_SYNCED, is_error=False).count())
+
+        # check admin token is redacted in HTTP logs
+        for log in HTTPLog.objects.all():
+            self.assertNotIn("WA_ADMIN_TOKEN", json.dumps(log.get_display()))
 
         mock_get.assert_called_with(
             "https://graph.facebook.com/v18.0/111111111111111/message_templates",
@@ -622,10 +628,9 @@ class WhatsAppTypeTest(TembaTest):
             headers={"Authorization": "Bearer WA_ADMIN_TOKEN"},
         )
 
-        # success no error and pagination
-        templates_data, no_error = WhatsAppType().get_api_templates(channel)
-        self.assertTrue(no_error)
-        self.assertEqual(["foo", "bar"], templates_data)
+        # check when templates across two pages
+        templates = channel.type.fetch_templates(channel)
+        self.assertEqual(["foo", "bar"], templates)
 
         mock_get.assert_has_calls(
             [